wakontrol/requirements.txt
Flask>=3.0.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
requests>=2.0.0
apscheduler>=3.10.0
